+++ drivers/tapdisk-log.h
@@ -58,7 +58,7 @@ void __tlog_error(const char *fmt, ...) __printf(1, 2);
 
 #define tlog_error(_err, _f, _a...)			\
 	__tlog_error("ERROR: errno %d at %s: " _f,	\
-		     _err, __func__, ##_a)
+		     (int)_err, __func__, ##_a)
 
 #define tlog_drv_error(_drv, _err, _f, _a ...) do {	\
 	if (tapdisk_driver_log_pass(_drv, __func__))	\
